+++ api/src/main/java/org/sigmapool/api/workers/models/WorkerResponseItem.kt
@@ -2,7 +2,7 @@ package org.sigmapool.api.workers.models
 
 class WorkerResponseItem(
     val title: String,
-    val hashrate: Long,
-    val hashrate24h: Long,
+    val hashrate: Float,
+    val hashrate24h: Float,
     val isOnline: Boolean
 )
